Add network metrics to Solaris PMDA
Use MIB2 interface available via TI to extract IPv4 interfaces and UDP stats
from the kernel.

The good/bad fetch is on the whole group - if any of the sources fail (even if
we don't use them like IPv6 interface stats) then whole fetch will return no
data.

Fetch callback still uses offset for the metrics which have indoms, for
singular metrics a pointer to the global storage is used whence the change
of type in md_offset.

+++ src/pmdas/solaris/common.h
@@ -73,7 +73,7 @@ int zfs_fetch(pmdaMetric *, int, pmAtomValue *);
 typedef struct {
     pmDesc	md_desc;	// PMDA's idea of the semantics
     int		md_method;	// specific kstat method
-    int		md_offset;	// offset into kstat stats structure
+    ptrdiff_t	md_offset;	// offset into kstat stats structure
 } metricdesc_t;
 
 extern metricdesc_t	metricdesc[];
